python_backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
google-generativeai
supabase
python-dotenv
langchain
langchain-community
langchain-text-splitters
langchain-google-genai
